Moved all Console arguments to a new namespace. Created SchedulerBootstrapper
