+++ Backend/Console/Commands/Domains/CurrencyDomainResolver.cs
@@ -1,4 +1,5 @@
-﻿using Console.Commands.Resolver.Assertions;
+﻿using Console.Commands.Domains.Arguments;
+using Console.Commands.Resolver.Assertions;
 using Console.Commands.Resolver.Pipelines;
 using Console.Types;
 using IdelPog.Common.Commands;
