+++ Backend/Console/Commands/Domains/PermissionDomainResolver.cs
@@ -1,4 +1,5 @@
-﻿using Console.Commands.Resolver.Assertions;
+﻿using Console.Commands.Domains.Arguments;
+using Console.Commands.Resolver.Assertions;
 using Console.Commands.Resolver.Pipelines;
 using Console.Runtime.Systems;
 using Console.Types;
