+++ Backend/Console/Commands/Resolver/Pipelines/CurrencyUpdateResolver.cs
@@ -1,4 +1,5 @@
-﻿using Console.Types;
+﻿using Console.Commands.Domains.Arguments;
+using Console.Types;
 using IdelPog.Common.Enums;
 
 namespace Console.Commands.Resolver.Pipelines
