+++ Backend/Scheduler/SchedulerBootstrapper.cs
@@ -1,0 +1,42 @@
+﻿using IdelPog.Common.Commands;
+using IdelPog.Common.Factories;
+using IdelPog.Messaging.Dispatch;
+using IdelPog.Messaging.Listeners;
+using IdelPog.Messaging.Messenger;
+using IdelPog.Messaging.Orchestration;
+using IdelPog.Validation.Assertions;
+using IdelPog.Validation.Assertions.Handlers;
+using IdelPog.Validation.Assertions.Handlers.Interfaces;
+using Scheduler.Controller;
+using Scheduler.Factory;
+using Scheduler.Mediator;
+using Scheduler.Register;
+using Scheduler.Runner;
+using Scheduler.Types;
+
+namespace Scheduler
+{
+    public class SchedulerBootstrapper
+    {
+        public void Initialize(IBufferMessenger bufferMessenger, IBufferManager bufferManager)
+        {
+            IHandler throwHandler = new ThrowHandler();
+            IAssertNonDuplicate assertNonDuplicate = new AssertNonDuplicate(throwHandler);
+            IAssertFound assertFound = new AssertFound(throwHandler);
+            IAssertNotNull assertNotNull = new AssertNotNull(throwHandler);
+            IAssertCollectionNotEmpty assertCollectionNotEmpty = new AssertCollectionNotEmpty(throwHandler);
+
+            IErrorDTOFactory errorDTOFactory = new ErrorDTOFactory();
+            ITaskErrorDTOFactory taskErrorDTOFactory = new TaskErrorDTOFactory(errorDTOFactory);
+            IDispatchOne<ScheduledTaskErrorDTO> errorDTODispatcher = new ManagedDispatcher<ScheduledTaskErrorDTO>(bufferManager, assertNotNull, assertCollectionNotEmpty);
+            IScheduleReader scheduleReader = new ScheduleRegister(assertNonDuplicate, assertFound, assertNotNull);   
+            IScheduleMediator scheduleMediator = new ScheduleMediator(scheduleReader, errorDTODispatcher, taskErrorDTOFactory, assertCollectionNotEmpty);
+            
+            IScheduleRunner scheduleRunner = new ScheduleRunner(scheduleMediator);
+            IScheduleController scheduleController = new ScheduleController(scheduleRunner);
+            ISingleListener<ScheduleControl> scheduleControlListener = new ScheduleControlListener(scheduleController);
+            
+            bufferMessenger.Subscribe(scheduleControlListener);
+        }
+    }
+}
